requirements.txt
pygame-ce
psutil
opencv-python
